LPS-179567 Extract CommercePricingClassDetailsScreenNavigationCategory methods to a CommercePricingClassDetailsScreenNavigationEntry component and extends CommercePricingClassDetailsScreenNavigationCategory in it
